refactor: extract .env loading into shared env_file module
将 gateway.rs 中的 .env 文件解析与加载逻辑提取到独立的 env_file 模块，
使 agent 和 gateway 命令共享同一套环境变量加载流程。agent.rs 现在在启
动时也会加载 blockcell .env 文件，确保两个入口命令行为一致。

变更概要：

- 新增 env_file.rs — 抽取共享的 .env 解析和加载逻辑，含单元测试
- 修改 gateway.rs — 移除已提取的函数，改为调用共享模块
- 修改 agent.rs — 在 run() 启动时新增 ensure_and_load_blockcell_env 调用
- 修改 mod.rs — 注册 env_file 模块
